refactor(core/query-builder): extract ColumnCondition into query-builder/condition (#405)
Pure move; re-exported from SelectQueryBuilder.ts.
